// src/app/models/dmo/dmo.ts
export class RbiAdviceForSaleOfSecurities {
    securityTypeId: string;
    currFaceValue: string;
    faceValSecSold: string;
    unitsSoldNo: string;
    salesPrice: string;
    saleSecRelzdAmt: string;
    brokenDays: string;
    accrIntrstRecv: string;
    totalRelzdAmt: string;
}